\name{mosaic.addFootnote}
\alias{mosaic.addFootnote}

\title{
addFootnote
}
\description{
Add a Footnote to plot using footnotestring and current timestamp.
}
\usage{
mosaic.addFootnote()
}

\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: internal
}

\keyword{ footnote }
\keyword{ timestamp }