\name{mosaic.beginPlot}
\alias{mosaic.beginPlot}
\title{
beginPlot
}
\description{
begin plotting the configured graphs for loaded data and generate the output PDF-File.
}
\usage{
mosaic.beginPlot(varname,outputfolder)
}
\arguments{
  \item{varname}{
name of the studyitem or  csv column loaded to plot graphs for.
}
\item{outputfolder}{
name of the output folder
}
}
\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: internal
}

\keyword{ plot }
