\name{mosaic.countValue}
\alias{mosaic.countValue}

\title{
countValue
}
\description{
Count occurrence of search value in data column
}
\usage{
mosaic.countValue(searchvalue, data_column)
}
\arguments{
  \item{searchvalue}{
value to search for
}
  \item{data_column}{
name of study item or data column to search in
}
}
\details{
useful to find qualified missings in data column
}
\value{
count of occurences of specified value in specified data column
}
\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: internal
}

\keyword{ count }
\keyword{ search }
