\name{mosaic.createSimplePdfCategoricalDataframe}
\alias{mosaic.createSimplePdfCategoricalDataframe}

\title{
createSimplePdfCategoricalDataframe
}
\description{
Create simple PDF-file for categorical data 
}
\usage{
mosaic.createSimplePdfCategoricalDataframe(df, outputfolder)
}

\arguments{
  \item{df}{
dataframe
}
  \item{outputfolder}{
path to output folder
}
}
\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: user
}




\keyword{ categorical }
\keyword{ pdf }
\keyword{ generate }
