\name{mosaic.generateCategoricalPlot}
\alias{mosaic.generateCategoricalPlot}

\title{
generateCategoricalPlot
}
\description{
Generate Statistics and Create plots for categorical data
}
\usage{
mosaic.generateCategoricalPlot(dataframe, varname)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dataframe}{
data table with one or more columns (first row should contain column names/study item names/variable names)
}
  \item{varname}{
selected column/study item/variable to plot graph for
}
}

\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: internal
}