\name{mosaic.generateMetricPlots}
\alias{mosaic.generateMetricPlots}

\title{
generateMetricPlots
}
\description{
calculate statistics and generate graphs for metric data
}
\usage{
mosaic.generateMetricPlots(data_snippet, var_name)
}

\arguments{
  \item{data_snippet}{
data table with one or more columns (first row should contain column names/study item names/variable names)
}
  \item{var_name}{
selected column/study item/variable to plot graph for
}
}

\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: internal
}