\name{mosaic.getTimestamp}
\alias{mosaic.getTimestamp}

\title{
getTimestamp
}
\description{
get a current timestamp formatted as \%Y_\%m_\%d_\%H\%M\%S
}
\usage{
mosaic.getTimestamp()
}
\value{
timestamp, e.g. '2016_09_09_143458'
}
\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: internal
}

\keyword{ timestamp }
