\name{mosaic.loadCsvData}
\alias{mosaic.loadCsvData}

\title{
loadCsvData
}
\description{
Load data from csv-file is one or more columns. first row should contain the name of the study item, e.g. 'height'
}
\usage{
mosaic.loadCsvData(filename)
}
\arguments{
  \item{filename}{
filename or a complete path to a file
}
}

\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: user
}

\keyword{ load }
\keyword{ csv }
\keyword{ data }