\name{mosaic.preProcessCategorialData}
\alias{mosaic.preProcessCategorialData}

\title{
preProcessCategorialData
}
\description{
Identify unique values in data column, get absolute, percentage and cumulative statistics
}
\usage{
mosaic.preProcessCategorialData(data)
}

\arguments{
  \item{data}{
data frame to be processed containing categorical data
}
}
\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: internal
}
