\name{mosaic.preProcessMetricData}
\alias{mosaic.preProcessMetricData}

\title{
preProcessMetricData
}
\description{
Pre-process metric data to allow missing-ratio table
}
\usage{
mosaic.preProcessMetricData(data)
}

\arguments{
  \item{data}{
data frame to be preprocessed containing metric data
}
}

\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: internal
}
