\name{mosaic.setGlobalDescription}
\alias{mosaic.setGlobalDescription}

\title{
setGlobalDescription
}
\description{
Set Global Description for variable User (description) data. especially useful when plotting graphs for a selected data column
}
\usage{
mosaic.setGlobalDescription(value)
}

\arguments{
  \item{value}{
string value to be used as study item description, e.g. 'waist circumference'
}
}

\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: user
}

\examples{
mosaic.setGlobalDescription('waist circumference')
}