\name{mosaic.setGlobalMissingTreshold}
\alias{mosaic.setGlobalMissingTreshold}

\title{
setGlobalMissingTreshold
}
\description{
Set Global Threshold for Missings , e.g. 99000
}
\usage{
mosaic.setGlobalMissingTreshold(value)
}

\arguments{
  \item{value}{
threshold to separate missings from valid values
}
}
\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: user
}


\examples{
mosaic.setGlobalMissingTreshold(99000)
}