\name{mosaic.setGlobalUnit}
\alias{mosaic.setGlobalUnit}
\title{
setGlobalUnit
}
\description{
Set Global Unit Label to be used User in graphs, e.g. '(cm)'
}
\usage{
mosaic.setGlobalUnit(value)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{value}{
unit string to be used in graphs
}
}

\author{
The MOSAIC Project, Martin Bialke
}

\note{
Function call type: user
}


\examples{
mosaic.setGlobalUnit('(cm)')
}
